Sistemato inserimento nel carrello e relativa servlet.
aggiunto doGet veloce nella servlet CartAdder. modificato metodo
addToCart(ItemInTheCart item) in Cart.java. Rimosso attributo "disabled"
dal tasto aggiungi al carrello di product.jsp (non serve a nulla)

+++ NibInk/src/main/java/com/model/Cart.java
@@ -6,21 +6,43 @@ public class Cart {
 	ArrayList<ItemInTheCart> cart; 
 	float total = 0;
 	
-	public Cart() {} 
+	public Cart() {
+		cart=new ArrayList<ItemInTheCart>();
+	} 
 	
-	public void addToCart(Item item, int quantity, String size)
+	public void addToCart(Item item, int quantity, String size) //per adesso inutilizzato
 	{
-		ItemInTheCart toAdd = new ItemInTheCart();
-		toAdd.setItem(item);
-		toAdd.setQuantity(quantity);
-		toAdd.setSize(size);
-		cart.add(toAdd);
+		boolean itemExists = false;
+		for(ItemInTheCart c : cart) {
+			System.out.println("item: "+c.getItem().getCodenumber() +"size: "+c.getSize() );
+			if(c.getItem().getCodenumber().equals(item.getCodenumber()) && c.getSize().equals(size)) {
+				c.setQuantity(c.getQuantity()+quantity);
+				itemExists = true;
+				break;
+			}
+		}
+		if(!itemExists) {
+			ItemInTheCart toAdd = new ItemInTheCart();
+			toAdd.setItem(item);
+			toAdd.setQuantity(quantity);
+			toAdd.setSize(size);
+			cart.add(toAdd);
+		}
 		total+=quantity*item.getPrice();
 	}
 	
 	public void addToCart(ItemInTheCart item)
 	{
-		cart.add(item);
+		boolean itemExists = false;
+		for(ItemInTheCart c : cart) {
+			if(c.getItem().getCodenumber().equals(item.getItem().getCodenumber()) && c.getSize().equals(item.getSize())) {
+				c.setQuantity(c.getQuantity()+item.getQuantity());
+				itemExists = true;
+				break;
+			}
+		}
+		if(!itemExists) 
+			cart.add(item);
 		total+=item.getQuantity()*item.getItem().getPrice();
 	}
 	
@@ -48,6 +70,10 @@ public class Cart {
 	public void setTotal(float newPrice)
 	{
 		this.total = newPrice;
+	}
+		
+	public ArrayList<ItemInTheCart> getCart() {
+		return cart;
 	}
 	
 }
